# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MediSafe Tracker API",
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "schemes": ["http"],
  "paths": {
    "/users/": {
      "get": {
        "summary": "Get all users (paginated)",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer" },
          { "name": "per_page", "in": "query", "type": "integer" }
        ],
        "responses": { "200": { "description": "List of users" } }
      },
      "post": {
        "summary": "Create a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "first_name": { "type": "string" },
                "last_name": { "type": "string" },
                "email": { "type": "string" }
              },
              "required": ["first_name", "last_name", "email"]
            }
          }
        ],
        "responses": {
          "201": { "description": "User created" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": {
          "200": { "description": "User found" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update a user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "integer" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "first_name": { "type": "string" },
                "last_name": { "type": "string" },
                "email": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "User updated" },
          "404": { "description": "User not found" },
          "409": { "description": "Email already exists" }
        }
      },
      "delete": {
        "summary": "Delete a user by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "description": "User not found" }
        }
      }
    },

    "/incidents/": {
      "get": {
        "summary": "Get all incidents (paginated)",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer" },
          { "name": "per_page", "in": "query", "type": "integer" }
        ],
        "responses": { "200": { "description": "List of incidents" } }
      },
      "post": {
        "summary": "Create a new incident",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "description": { "type": "string" },
                "severity": { "type": "string" },
                "user_id": { "type": "integer" }
              },
              "required": ["title", "description", "severity", "user_id"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Incident created" },
          "400": { "description": "Invalid data or constraint violation" }
        }
      }
    },
    "/incidents/{id}": {
      "get": {
        "summary": "Get an incident by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": {
          "200": { "description": "Incident found" },
          "404": { "description": "Incident not found" }
        }
      },
      "put": {
        "summary": "Update an incident by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "integer" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "description": { "type": "string" },
                "severity": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Incident updated" },
          "404": { "description": "Incident not found" },
          "400": { "description": "Invalid update" }
        }
      },
      "delete": {
        "summary": "Delete an incident by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": {
          "200": { "description": "Incident deleted" },
          "404": { "description": "Incident not found" }
        }
      }
    }
  }
}